refactor: changed symbols structure, using new upper class 'Symbol'. Added class SubjektFunction instead of using Kotlin Function1
